+++ scripts/menus/programs/program-vpn.sh
@@ -34,7 +34,7 @@ case $CHOICE in
      "2)")
       ansible-playbook /opt/plexguide/ansible/vpn.yml --tags rtorrentvpn
       echo "RTorrentVPN: http://$ipv4:3000"
-      echo "For Subdomain http://rtorrentvpn.$domain"
+      echo "For Subdomain https://rtorrentvpn.$domain"
       echo "For Domain http://$domain:3000"
       echo ""
       echo "Please set your own username & password!"
@@ -45,7 +45,7 @@ case $CHOICE in
      "3)")
       ansible-playbook /opt/plexguide/ansible/vpn.yml --tags delugevpn
       echo "DelugeVPN: http://$ipv4:8112"
-      echo "For Subdomain http://delugevpn.$domain"
+      echo "For Subdomain https://delugevpn.$domain"
       echo "For Domain http://$domain:8112"
       echo ""
       echo "Default password: deluge"